requirements.txt
fastapi
uvicorn[standard]
zigpy
bellows
aiomqtt
pyaml
zha-quirks
duckdb